requirements.txt
fastapi
pydantic
pydantic-settings
pydantic[email]
uvicorn
passlib
python-jose
pymongo
bcrypt

google-generativeai
numpy